refactor: rewrite Forms component to use hooks
