+++ src/components/Form/Form.spec.tsx
@@ -2,7 +2,8 @@ import { expect, test } from "vitest";
 import { render, screen } from "@testing-library/react";
 import Form from "@/components/Form/Form";
 import React from "react";
-import { IFormCard } from "@/pages/Forms/Forms";
+
+import { IFormCard } from "@/interfaces/formCard.interface";
 
 const mockOnSubmit: (cards: IFormCard) => void = () => {};
