+++ src/interfaces/cardBorderColor.type.ts
@@ -1,0 +1,1 @@
+export type TCardBorderColor = "red" | "black" | "blue";
